requirements.txt
Flask
Cython
grpcio
matplotlib
numpy
opencv-python
pandas
pillow
#pycocotools
requests
scikit-image
#shutil
tensorflow==1.15